+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.x.x] - 2018-XX-XX
+## [0.7.11] - 2018-06-11
 ### Added
 - Push Notifications
 - Add deep linking